refacored according to pr comments
